pyproject.toml
[project]
name = "ai-engineer"
version = "0.1.0"
description = "ai-engineering apps"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=6.5.1",
    "numpy>=2.4.2",
    "openai>=2.17.0",
    "openai-agents>=0.8.3",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.7.0",
    "requests>=2.32.5",
    "rich>=14.3.2",
]
